+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/report/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package report contains everything needed to build a
+ * {@link org.hisp.dhis.tracker.report.TrackerImportReport} which we return at
+ * the end of the tracker import process.
+ */
+package org.hisp.dhis.tracker.report;
