+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/ValidationErrorReporter.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.tracker.report;
+package org.hisp.dhis.tracker.validation;
 
 import java.util.ArrayList;
 import java.util.EnumMap;
@@ -42,11 +42,18 @@ import lombok.Value;
 import org.hisp.dhis.tracker.TrackerIdSchemeParams;
 import org.hisp.dhis.tracker.TrackerType;
 import org.hisp.dhis.tracker.domain.TrackerDto;
-import org.hisp.dhis.tracker.validation.ValidationFailFastException;
+import org.hisp.dhis.tracker.report.TrackerErrorCode;
+import org.hisp.dhis.tracker.report.TrackerErrorReport;
+import org.hisp.dhis.tracker.report.TrackerWarningReport;
 
 /**
- * A class that collects {@link TrackerErrorReport} during the validation
- * process.
+ * Collects {@link TrackerErrorReport}s, {@link TrackerWarningReport}s and
+ * invalid entities the errors are attributed to.
+ * <p>
+ * Long-term we would want to remove the responsibility of tracking invalid
+ * entities from here. This could allow us to merge this class with
+ * {@link org.hisp.dhis.tracker.report.TrackerValidationReport}.
+ * </p>
  *
  * @author Morten Svanæs <msvanaes@dhis2.org>
  */
@@ -61,7 +68,7 @@ public class ValidationErrorReporter
 
     TrackerIdSchemeParams idSchemes;
 
-    @Getter( AccessLevel.NONE )
+    @Getter( AccessLevel.PACKAGE )
     /*
      * Keeps track of all the invalid Tracker objects (i.e. objects with at
      * least one TrackerErrorReport in the ValidationErrorReporter) encountered
